G8-39 : updated code to process dublin bus response

+++ external_data/src/main/java/com/tcd/ase/externaldata/model/dublinBus/Header.java
@@ -15,5 +15,5 @@ public class Header {
     public String incrementality;
 
     @JsonProperty("Timestamp")
-    public int timestamp;
+    public Double timestamp;
 }
